app/trained model/api_service/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
pillow>=10.0.0
numpy>=1.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
tensorflow-cpu>=2.20.0
requests>=2.30.0
